Move components to Pages folder
