+++ src/pages/Home/Home.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Sidebar from '../Sidebar/Sidebar';
-import CarDetails from '../CarDetails/CarDetails';
+import Sidebar from '../../components/Sidebar/Sidebar';
+import CarDetails from '../../components/CarDetails/CarDetails';
 
 const Home = () => (
   <div className="d-flex justify-content-between align-items-center background-container vw-100">
